Include the client-side CLI utils in the image
The image can now be used to run an ldap client command rather than the
slapd server. If so, skip slapd-specific configuration and setup logic
and do not require the environment variables used by this setup. This
approach is inspired by other Docker entrypoint scripts used for initial
setup such as in the Docker Hub-official MySQL and Postgres images, but
uses an alternative exit point for the script instead of an if clause
that spans most of the logic.